requirements.txt
# Graph-based Federated Learning for Malware Detection
# Research-grade dependencies for IIT Roorkee server

# Core Deep Learning
torch>=1.12.0
torch-geometric>=2.1.0
torch-scatter>=2.0.9
torch-sparse>=0.6.15
torch-cluster>=1.6.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.8.0
scikit-learn>=1.1.0

# Data Processing
pandas>=1.4.0
networkx>=2.8.0

# Configuration & Logging
pyyaml>=6.0
tqdm>=4.64.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Development & Testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0